Move CompilationView to a separate file
